feat: `diff` shows moved resources (#708)
When a resource is moved from one location to another, the `diff`
command will show two entries: an addition and a removal. This PR
augments those entries to show where they moved to and from,
respectively:

<img width="1440" height="709" alt="Drawing 2025-07-15 14 35 02
excalidraw"
src="https://github.com/user-attachments/assets/005c27a0-b433-4203-a3ee-7bfa893d3f47"
/>


This information is not shown by default. To enable it, the user has to
pass `--include-moves` to the CLI, and `includeMoves: true` to the
toolkit.

Closes https://github.com/aws/aws-cdk-cli/issues/362

Depends on https://github.com/aws/aws-cdk-cli/pull/679.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>