fix(integ-runner): error when running the update workflow with toolkit-lib engine (#717)
Fixes #715

The update workflow will deploy an existing snapshot from the main
branch of the project before attempting to update the stack with
incoming changes.

When the toolkit-lib engine was used, the snapshot directory was
incorrectly treated as an CDK app command and failing.

The fix is to check if the app value is a directory and if so, just use
it as a cloud assembly directory.


---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license