chore(toolkit-lib): cleanup unused  and private refactor exports (#722)
Removes unused and internal exports for refactor. `MappingSource` is not
used anymore for now. Mark `RefactorOptions` as `readonly` as per our
best practices. Move `mappingsByEnvironment` and `parseMappingGroups` to
be internal as intended.

Also makes the API Extractor task fail on errors now, to prevent future
issues with missing exports. This is how I noticed the unintentional
exports in the first place.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license