fix(integ-runner): command to reproduce snapshot checkout does not always work (#727)
During the update workflow, when checking out the previous snapshot
failed, we print a git command to allow the user to reproduce potential
issues. This command doesn't work as printed, as it assumes it is run
from the same directory the test case file is in. However in most
situations a users will have run `integ-runner` from a parent directory.

It would print something like this, assuming you are in the same
directory as the test file:
```console
git checkout $(git merge-base HEAD main) -- integ.bucket-deployment-deployed-bucket.js.snapshot
```

The change in this PR updates `git` operations to use the `-C` flag to
specify the correct directory instead of changing the working directory
for the shell execution. Also changes the base the directory the
snapshot is in, instead of `this.directory`. This results in the same
dir, but is more localized to the feature and does not rely on the fact
that `this.directory` and `this.snapshotDir` happen to share the same
parent.

So now, we print a working command, where ever you are:
```console
git -C path/to/tests checkout $(git -C path/to/tests merge-base HEAD main) -- integ.bucket-deployment-deployed-bucket.js.snapshot
```

I could have just changed the printed command, but decided that
consistency between what we run and what we show is important. This also
simplifies the code because we don't have to manually maintain the
logger output anymore.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license