fix(cli): running `--telemetry-file` without the `--unstable` option creates a telemetry file (#731)
This feature is currently unstable, so it shouldn't apply at all if
invoked without the `--unstable` flag.

Right now, it creates a telemetry file with an event indicating the
aforementioned failure.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license