chore(cli): allow valid enum values in telemetry data (#711)
Enum values are defined as `choices` in the `cli-type-registry.json`
file. If the choices array exists, and if the value we get is included
as a choice, then we allow that raw value to be recorded in telemetry
because it is not free text. Invalid values passed in an enum property
are not recorded. In practice, the CLI fails before telemetry starts
with

```
Invalid values:
  Argument: require-approval, Given: "neveasdfr", Choices: "never", "any-change", "broadening"
```

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>