test(cli-integ): output go version in init-go integration tests (#766)
This change adds a `go version` check to the init-go integration test to
ensure Go is available and allow checking the version before running the
CDK init command.

This helps provide better debugging information if the test fails due to
Go not being available in the test environment.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license