chore(cli): funnel (#760)
A funnel is a combination sink class. It will be used later to record
telemetry both to file _and_ send to endpoint. That version will be used
in the situation where telemetry is turned on _and_ `--telemetry-file`
is set.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>