fix: cdk flags hint is shown before template payload (#785)
When a user runs `cdk synth`, they receive a message about their
unconfigured feature flags. This relocates the message to be displayed
after the CloudFormation templates, to make it more visible.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license