feat: allow filtering `cdk flags` feature flags by substring (#776)
This allows users to run `cdk flags` with a substring and all matching
feature flags will be displayed to them.

When a user runs `cdk flags #substring#` and there is only one matching
flag, a description of the flag will be displayed. If there are multiple
matching flags, all matching flags will be displayed in a table.

If a user runs `cdk flags` with multiple substrings, all matching flags
that contain any of the substrings will be displayed in a table.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license