fix(cli): deprecate "bootstrap-stack-name" in favor of "toolkit-stack-name" for gc command (#795)
Fixes #385

Elsewhere we are consistently naming this property `toolkit-stack-name`,
so this PR aligns `gc` with the rest of the CLI world. We may have
better options here, like making `toolkit-stack-name` a global option,
but this is an easy fix for now.

Does not break users by maintaining `bootstrap-stack-name` as an alias
of `toolkit-stack-name` for `gc`. However `bootstrap-stack-name` is
marked as deprecated and will be removed when we GA garbage collection.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license