fix: skipLibCheck is not turned on for all init templates (#800)
This makes TypeScript check whether all provided type files are
coherent, which is taking more and more time as the CDK is growing.

Turning this off greatly decreases type checking and compilation times.

See:
https://devblogs.microsoft.com/typescript/announcing-typescript-5-9/#minimal-and-updated-tsc---init

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license