fix(cli): flags command shows unnecessary flags, not all table rows have vertical lines (#796)
When a user runs `cdk flags`, a table is displayed with information
about their feature flags. This PR omits the intentionally unconfigured
feature flags from the table and makes the table more visually
appealing.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

---------

Co-authored-by: Rico Hermans <rix0rrr@gmail.com>