fix(cli): fails to load credentials if both containerToken and containerTokenFile are set (#829)
If both `$AWS_CONTAINER_AUTHORIZATION_TOKEN` and
`$AWS_CONTAINER_AUTHORIZATION_TOKEN_FILE` are set, the SDK wants to emit
a warning.

Because of the way it calls the `logger.warn()` function, the reference
to `this` is lost and we try to look up a member on `undefined` which
then fails and no credentials are loaded at all.

Reported as an upstream bug to SDKv3, but in order to expedite the fix
we also guard against it ourselvses by locally binding `this` before we
pass our logger object to the SDK.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license