fix: notices key in cdk.json is not respected (#821)
Fixes #724

## Description
Setting `"notices": false` in cdk.json doesn't suppress notices.

The root cause is that we handle logic to determine whether we should
print notices at *yargs*. Yargs had a default value determined by a
helper function (`YARGS_HELPERS.shouldDisplayNotices()`) that returned
true when not in CI, or when in a safe CI.

This meant that any setting in our configuration would be overwritten by
the output of this helper, because it looks like it's coming from the
command line!

## Solution

Centralize logic for notices in `cli.ts`.

Made the notices option in yargs `undefined` by default. Moved the logic
from `YARGS_HELPERS.shouldDisplayNotices()` into the `cli.ts`. This has
several implications.
- Setting `notices` in config now functions as expected; it will still
not override the command line setting
- If a CDK app is running in a CI environment where it would not be safe
to print notices
- NEW BEHAVIOR: Notices WILL be printed if `notices: true` or `notices:
[truthy value]` is set in config
- Notices WILL be printed if flag `--notices` is passed, and NOT printed
if `--no-notices` / `--notices=false` is passed at the command line (was
already the case)

**Caveat: `"notices": "false"` in the config will suppress notices, but
other "truthy" values will cause them to be printed.**

### Testing
Added unit tests to verify the hierarchy of behavior:
1. Flags in CLI (`--notices` or `--no-notices`)
2. Setting in configuration (`"notices": false` in cdk.json)
3. Print notices if not in CI or in CI where it is safe

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>