chore: improve stability of tests by copying SSM Parameter Value (#830)
`{Fn::GetAtt}` on an `AWS::SSM::Parameter` is eventually consistent. If
an SSM Parameter was just created in a stack deployment, the
`{Fn::GetAtt}` may sometimes fail with a "Parameter not found" error.

This causes ~2 canary failures per week (about 1% failure rate).

We don't actually need to `{Fn::GetAtt}` the value as we know what it is
(it is static, after all). Copy the same literal value into both places
to reduce the canary failure rate by a little bit, and add a test to
make sure the values don't accidentally drift apart.

Internal reference D259904064.


---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license