chore(cli-lib-alpha): stop releasing deprecated package (#905)
Stop releasing the deprecated `@aws-cdk/cli-lib-alpha` package. This
change removes the package from the code base and integ tests. If we
ever need to release it again, we can revert this PR.

I tried to just disable releasing, but then integ tests fail and I think
removing it all in one PR is just easier if we need to go back to it.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license