feat(cli-integ): release @aws-cdk-testing/cli-integ (#912)
#905 intended to trigger a release for `@aws-cdk-testing/cli-integ`,
where it did not rely on the deprecated `cli-lib-alpha` module. because
it was marked as a chore, the release was skipped and the latest version
still runs the tests that were removed in #905. These tests are now
failing because they cannot find the right version of `cli-lib-alpha`.
By releasing a version of `@aws-cdk-testing/cli-integ`, we won't have to
worry about this anymore.

this PR is intended to trigger a release and nothing else.


---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license