feat(cli): display warning when --role-arn is used with gc command (#893)
The gc command currently ignores the --role-arn parameter and always
uses the user's default credentials, regardless of whether a role ARN is
specified. This can potentially lead to silent failures, permission
issues, as well as a bad user experience due to confusion.

This pr adds a check to throw a warning when the --role-arn parameter is
used in the gc command. This helps improve user experience and avoid
mismanaged expectations.

A consideration to throw an error was made. However, this could be a
breaking change for existing ci/cd pipeline users. Therefore, this idea
was discarded

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license