fix(integ-runner): inconsistent and unexpected CWD in test app execution (#720)
Fixes an issue with the working directory being inconsistent and
unexpected during execution of the test app. Previously we had two
different behaviors: For regular execution we would execute the app from
the directory of the test file, but for watch we would run it in the
actual user's CWD. This is inconsistent behavior making it difficult to
write integ test cases than can be executed across all modes.

Additionally as reported in #638, languages likes python may rely on the
working directory to import modules. The working directory changing
compared to library code and modes, makes it very difficult to use
integ-runner in these languages.

This change aligns the working directory for test app execution to
always be the CWD integ-runner is executed from. Given the two existing
choices, this is the less arbitrary one and aligns more with how running
a test app directly (`node integ.app-under-test.test.js`) would work.

The practical implications of this are limited. Tests are already
discovered and executed using relative paths. In case someone really
dependent on this behavior, they can simply execute integ-runner from a
different working dir.

Fixes #638

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license