fix(aws-cdk): init templates use outdate node types (#941)
Fixes https://github.com/aws/aws-cdk-cli/issues/935

The app needs to be on node@24, as we would like customers to run their
applications on latest LTS of Node.js.
The lib needs to be node@20, as we would like customers to support
minimum number of supported Node.js version.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

---------

Co-authored-by: Momo Kornher <mail@moritzkornher.de>