feat(toolkit-lib): network detector (#926)
this PR implements a connectivity detector that performs a HEAD request
to the notices endpoint, with a timeout of 1 hour. before making any
subsequent network calls, (including GET notices and PUT telemetry) we
check to see if we have already determined the network to be
disconnected. in those cases, we skip subsequent calls.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>