fix(integ-runner): region is not passed to CDK App (#949)
When using the `toolkit-lib` deployment engine, the selected region was
not being passed to the CDK app.

In the old engine, we could set the `AWS_REGION` environment variable
and the CDK CLI would respect that, but the toolkit lib was passing that
environment variable directly to the CDK app which was not respecting
that. Fix that by threading the region through as an explicit parameter
and adding an override parameter to the `AwsCliCompatible` base
credentials: we don't necessarily want to read this only from the INI
files.

Additionally: add a validation to make sure that the CDK app respects
the region we are requesting. CDK integ tests should either not pass an
environment, or read `$CDK_DEFAULT_REGION` to determine the current
region. It was too easy to just override the region with a hardcoded
one, but that breaks the integ runner's attempts to isolate tests across
regions. Guard against that by validating.

Also in this PR: disable the update workflow. I've lost 2 hours
debugging why my deployment was failing with nonsenical values and the
answer was: the checked-in snapshot was nonsensical, and the update
workflow will first deploy the checked-in snapshot and only then deploy
the current snapshot. Good idea, bad execution: if existing snapshots
are not region-agnostic they are guaranteed to fail. We should probably
resynthesize them from the previous git commit instead. Whatever it is,
what we do right now doesn't work so I'm disabling it.

Fixes #

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license