chore(cli-integ): run telemetry integ tests separately (#962)
this is done to avoid IP address throttling. in #775, these tests are
modified to explicitly check if telemetry is sent successfully, so we
must make sure that they are not throttled.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license