feat: show early validation errors on deploy (#970)
CloudFormation has [recently launched early validation][1]. When
creating the change set, CloudFormation validates a few things such as
whether the deploy would try to create resources with existing physical
IDs. More validations to come. But the specific information about the
error (other than it's an early validation error) is not returned by the
`DescribeChangeSet` API. The consumer is supposed to call a new API,
`DescribeEvents` to get that information.

Introduce a new class,`EarlyValidationReporter` that checks whether such
an error occurred and call the `DescribeEvents` API. The bootstrap stack
was also updated to include permission to this new API.

If the role being used for deployment doesn't have permission to call
`DescribeEvents`, the toolkit will warn the user that it cannot show the
proper error message, and suggest a re-bootstrap.


[1]:
https://aws.amazon.com/about-aws/whats-new/2025/11/cloudformation-dev-test-cycle-validation-troubleshooting/

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license