chore: upgrade SDK and remove process.exit (#981)
SDK version v3.946.0 has the fix to the metadata service lookup hanging,
so upgrade and remove the explicit `process.exit()`.

The dependency upgrade surfaced a bug in `yarn-cling` which is being
fixed here as well.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

---------

Signed-off-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>