feat: init-templates enable test cleanup by default (#986)
Since `aws-cdk-lib` 2.232.0, it has support for enabling automatic
temporary test directory cleanup in jest:
https://github.com/aws/aws-cdk/pull/36226 This needs to be explicitly
enabled via jest config though.

Upgrade to the latest `aws-cdk-lib` version and turn on that setting in
the init templates.


---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

---------

Signed-off-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>