fix(integ-runner): `--profile` not working with toolkit-lib engine (#996)
Fixes https://github.com/aws/aws-cdk-cli/issues/993

This change adds `profile` support to the toolkit-lib engine in the
integ-runner package. The profile option is now passed through from the
IntegRunnerOptions to the ToolkitLibEngineOptions and ultimately to the
BaseCredentials configuration. This allows users to specify which AWS
profile to use when running integration tests with the toolkit-lib
engine, providing consistency with other authentication options already
available in the integ-runner.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license