fix(cli): error message when early validation failures cannot be retrieved is unclear (#995)
Clarified error message for early validation failures in deployments.

The current error message is confusing: it can be taken to mean that the
CLI version is outdated, which is not the case.

It also mentions the CDK Toolkit stack version without any explanation -
it is not clear which version it is referring to.

I also believe that saying "with the latest version of the CLI" is
redundant - the error message will only appear in a sufficiently
up-to-date CLI.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license