fix(cli): upgrade jest in init-templates to remove one deprecation warning (#1009)
Via a long transitive dependency chain, jest 29 depends on a package
called `inflight` that throws up a warning on install:

```
npm warn deprecated inflight@1.0.6: This module is not supported, and leaks memory. Do not use it. Check out lru-cache if you want a good and tested way to coalesce async requests by a key value, which is much more comprehensive and powerful.
```

Upgrade the package.json we generate upon `cdk init` to jest 30 so that
customers don't see this warning. Even though everything works fine and
this module is being used for testing so even if it leaked it wouldn't
matter, it still gives them bad vibes.

We still have the following installation warning:

```
npm warn deprecated glob@7.2.3: Glob versions prior to v9 are no longer supported
```

This dependency comes via:

```
jest 30.2 -> @jest/transform 30.2 -> babel-plugin-instanbul 7.0.1 -> test-exclude 6.0 -> glob 7.2.3
```

So still present in the latest jest, and there's nothing we can do about
it.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license